% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{book1.100k}
\alias{book1.100k}
\title{Example Dataset: book1_100k}
\format{
A data frame with the following columns:
\describe{
  \item{Id}{Unique identifier for the book.}
  \item{Name}{The title of the book.}
  \item{RatingDist1, RatingDist2, ..., RatingDist5}{Distribution of user ratings.}
  \item{pagesNumber}{The number of pages in the book.}
  \item{CountsOfReview}{The number of reviews.}
  \item{Rating}{The overall rating of the book.}
  \item{PublishYear, PublishMonth, PublishDay}{Publication date details.}
  \item{Authors}{The authors of the book.}
}
}
\source{
User-provided dataset
}
\usage{
book1.100k
}
\description{
This dataset contains metadata for books with 58,293 observations and 18 variables.
}
\keyword{datasets}
