% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_books.R
\name{recommend_books}
\alias{recommend_books}
\title{Recommend Similar Books}
\usage{
recommend_books(book_name, book_features, similarity_matrix, top_n = 5)
}
\arguments{
\item{book_name}{The name of the book for which recommendations are needed.}

\item{book_features}{A data frame containing book information.}

\item{similarity_matrix}{A similarity matrix between books.}

\item{top_n}{Number of recommendations to return.}
}
\value{
A data frame of recommended books.
}
\description{
This function recommends books similar to a given book.
}
